+++ test/layers/VerticalLineLayer-test.ts
@@ -3,9 +3,9 @@ import { expect } from 'chai';
 
 import { method } from './layerTestUtils';
 import CanvasContextSpy from '../../src/test-util/CanvasContextSpy';
-import { _renderCanvas } from '../../src/core/layers/HoverLineLayer';
+import { _renderCanvas } from '../../src/core/layers/VerticalLineLayer';
 
-describe('HoverLineLayer', () => {
+describe('VerticalLineLayer', () => {
   let spy: typeof CanvasContextSpy;
 
   const DEFAULT_PROPS = {
@@ -17,41 +17,41 @@ describe('HoverLineLayer', () => {
     spy = new CanvasContextSpy();
   });
 
-  function renderWithSpy(spy: CanvasRenderingContext2D, hover?: number) {
-    _renderCanvas(_.defaults({ hover }, DEFAULT_PROPS), 100, 100, spy);
+  function renderWithSpy(spy: CanvasRenderingContext2D, xValue?: number) {
+    _renderCanvas(_.defaults({ xValue }, DEFAULT_PROPS), 100, 100, spy);
   }
 
-  it('should do nothing if no hover value is provided', () => {
+  it('should do nothing if no xValue is provided', () => {
     renderWithSpy(spy, undefined);
 
     expect(spy.operations).to.deep.equal([]);
   });
 
-  it('should do nothing if the hover value is NaN', () => {
+  it('should do nothing if the xValue is NaN', () => {
     renderWithSpy(spy, NaN);
 
     expect(spy.operations).to.deep.equal([]);
   });
 
-  it('should do nothing if the hover value is infinite', () => {
+  it('should do nothing if the xValue is infinite', () => {
     renderWithSpy(spy, Infinity);
 
     expect(spy.operations).to.deep.equal([]);
   });
 
-  it('should do nothing if the hover value is before the X domain', () => {
+  it('should do nothing if the xValue is before the X domain', () => {
     renderWithSpy(spy, -100);
 
     expect(spy.operations).to.deep.equal([]);
   });
 
-  it('should do nothing if the hover value is after the X domain', () => {
+  it('should do nothing if the xValue is after the X domain', () => {
     renderWithSpy(spy, 200);
 
     expect(spy.operations).to.deep.equal([]);
   });
 
-  it('should render a hover line for a hover value in bounds', () => {
+  it('should render a value line for a xValue in bounds', () => {
     renderWithSpy(spy, 50);
 
     expect(spy.callsOnly('moveTo', 'lineTo')).to.deep.equal([
@@ -60,7 +60,7 @@ describe('HoverLineLayer', () => {
     ]);
   });
 
-  it('should round the hover value to the integer', () => {
+  it('should round the xValue to the integer', () => {
     renderWithSpy(spy, 33.4);
 
     expect(spy.callsOnly('moveTo', 'lineTo')).to.deep.equal([