+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `SimpleLineLayer` renamed to `LineLayer`. [#87](https://github.com/palantir/react-layered-chart/issues/87).
 - `SpanDatum` renamed to `BarDatum`, `XSpanDatum` to `SpanDatum`. [#48](https://github.com/palantir/react-layered-chart/issues/48).
 - `DataLoader`'s signature has had its signature rewritten to remove redundant arguments. [#82](https://github.com/palantir/react-layered-chart/issues/82).
+- `HoverLineLayer` renamed to `VerticalLineLayer` and had its `hover` prop renamed to `xValue`. [#119](https://github.com/palantir/react-layered-chart/issues/119).
 
 ### Fixed